+++ test-suite/bugs/closed/bug_3682.v
@@ -1,6 +1,6 @@
 Require Import TestSuite.admit.
 Class Foo.
 Definition bar `{Foo} (x : Set) := Set.
-Instance: Foo.
+Instance: Foo := {}.
 Definition bar1 := bar nat.
 Definition bar2 := bar ltac:(admit).