+++ theories/Classes/SetoidTactics.v
@@ -41,6 +41,7 @@ Definition default_relation `{DefaultRelation A R} := R.
   (lowest priority). *)
 
 Instance equivalence_default `(Equivalence A R) : DefaultRelation R | 4.
+Defined.
 
 (** The setoid_replace tactics in Ltac, defined in terms of default relations
   and the setoid_rewrite tactic. *)